desktop/requirements_desktop.txt
# Desktop application requirements
# Includes all base requirements plus desktop-specific dependencies

# Base requirements (relative to desktop directory)
-r ../requirements.txt

# Desktop packaging
pyinstaller>=6.0.0

# System tray support (optional but recommended)
pystray>=0.19.5
pillow>=10.0.0

# HTTP client for health checks
requests>=2.31.0
